group + online user

+++ src/controllers/groups.controller.js
@@ -1,6 +1,8 @@
 import Group from "../models/group.model.js";
 import GroupMember from "../models/group_members.model.js";
 import mongoose from "mongoose";
+import User from "../models/users.model.js";
+import Message from "../models/messages.model.js";
 // Tạo nhóm chat riêng tư giữa hai người dùng
 // Hàm này sẽ được gọi khi người dùng muốn tạo nhóm chat riêng tư với một người bạn
 const createPrivateGroup = async (req, res) => {
@@ -9,7 +11,8 @@ const createPrivateGroup = async (req, res) => {
     const userId = req.user._id;
     // Chuyển sang kiểu objectid trong mongoose
     const objectUserId = mongoose.Types.ObjectId.createFromHexString(userId);
-    const objectFriendId = mongoose.Types.ObjectId.createFromHexString(friendId);
+    const objectFriendId =
+      mongoose.Types.ObjectId.createFromHexString(friendId);
 
     // Kiểm tra xem nhóm hai người đã tồn tại chưa
     const privateGroup = await Group.findOne({
@@ -28,13 +31,17 @@ const createPrivateGroup = async (req, res) => {
         ]).then((res) => res.map((r) => r._id)),
       },
     });
-    
+
     if (privateGroup) {
-      return res.status(200).json({ message: "Nhóm hai người đã tồn tại", group: privateGroup });
+      return res
+        .status(200)
+        .json({ message: "Nhóm hai người đã tồn tại", group: privateGroup });
     }
-    
+
+    const friend = await User.findById(friendId);
+
     // Tạo nhóm mới
-    const group = new Group({ name: null, type: "private" });
+    const group = new Group({ name: friend._id, type: "private" });
     await group.save();
 
     // Thêm người dùng hiện tại vào nhóm
@@ -98,8 +105,25 @@ const createGroupChat = async (req, res) => {
 // Hàm này sẽ được gọi khi người dùng muốn xóa một thành viên khỏi nhóm
 const removeMemberFromGroup = async (req, res) => {
   try {
-    const { groupId, memberId } = req.body;
+    const groupId = req.params.groupId;
+    const { memberId } = req.body;
     const userId = req.user._id; // ID của người dùng hiện tại
+    const groupMember = await GroupMember.findOne({
+      group_id: groupId,
+      user_id: userId,
+      role: "admin",
+    });
+    if (!groupMember) {
+      return res
+        .status(403)
+        .json({ message: "Bạn không có quyền xóa thành viên khỏi nhóm" });
+    }
+
+    if(groupMember.user_id.toString() === memberId) {
+      return res
+        .status(403)
+        .json({ message: "Bạn không thể xóa chính mình khỏi nhóm" });
+    }
     // Xóa thành viên khỏi nhóm
     const result = await GroupMember.findOneAndDelete({
       group_id: groupId,
@@ -118,18 +142,83 @@ const removeMemberFromGroup = async (req, res) => {
 };
 
 // Lấy danh sách nhom mà người dùng tham gia
-// Hàm này sẽ được gọi khi người dùng muốn xem danh sách nhóm mà họ tham gia
+// Hàm này sẽ được gọi khi người dùng muốn xem tất cả danh sách nhóm mà họ tham gia
 const getUserGroups = async (req, res) => {
   try {
     const userId = req.user._id;
+    // Lấy danh sách nhóm người dùng hiện tại tham gia
 
-    // Lấy danh sách nhóm mà người dùng tham gia
     const groups = await GroupMember.find({ user_id: userId }).populate(
       "group_id",
       "name type created_at"
     );
 
-    res.status(200).json({ message: "Danh sách nhóm", groups });
+    const results = [];
+
+    for (const group of groups) {
+      const groupId = group.group_id._id;
+
+      // Lấy tất cả thành viên của nhóm
+      const members = await GroupMember.find({ group_id: groupId });
+
+      if (group.group_id.type === "private") {
+        // Nếu là nhóm riêng tư, lấy thông tin của bạn bè
+        const friendId = members.find(
+          (member) => member.user_id.toString() !== userId
+        ).user_id;
+        const friend = await User.findById(friendId);
+        const friendInfo = {
+          _id: friend._id,
+          username: friend.username,
+          avatar: friend.avatar,
+        };
+
+        const lastMessage = await Message.findOne({ group_id: groupId })
+          .sort({ created_at: -1 })
+          .select("content sender_id created_at");
+
+        results.push({
+          _id: groupId,
+          name: friend.username,
+          created_at: group.group_id.created_at,
+          type: group.group_id.type,
+          active_avatar: friendInfo.avatar,
+          last_message: lastMessage || null,
+        });
+      } else if (group.group_id.type === "group") {
+        const lastMessage = await Message.findOne({ group_id: groupId })
+          .sort({ created_at: -1 })
+          .select("content sender_id created_at");
+
+        results.push({
+          _id: groupId,
+          name: group.group_id.name,
+          created_at: group.group_id.created_at,
+          type: group.group_id.type,
+          active_avatar: "null",
+          last_message: lastMessage || null,
+        });
+      }
+    }
+
+    res.status(200).json({ message: "Danh sách nhóm", groups: results });
+  } catch (err) {
+    res.status(500).json({ message: err.message });
+  }
+};
+
+// Lấy danh dách nhóm public
+// Hàm này sẽ được gọi khi người dùng muốn xem danh sách nhóm công khai
+const getJoinedGroups = async (req, res) => {
+  const userId = req.user._id;
+  try {
+    const memberships = await GroupMember.find({ user_id: userId }).populate({
+      path: "group_id",
+      match: { type: "group" }, // chỉ lấy nhóm công khai
+    });
+    const groups = memberships.map((m) => m.group_id).filter((g) => g !== null); // loại bỏ null do không match
+
+    res.status(200).json({ message: "Danh sách nhóm đã tham gia", groups });
   } catch (err) {
     res.status(500).json({ message: err.message });
   }
@@ -141,23 +230,31 @@ const getGroupMembers = async (req, res) => {
   try {
     const { groupId } = req.params;
 
-    // Lấy danh sách thành viên trong nhóm
-    const members = await GroupMember.find({ group_id: groupId }).populate(
-      "user_id",
-      "username email"
-    );
+    // Lấy danh sách thành viên trong nhóm và thông tin người dùng
+    const members = await GroupMember.find({ group_id: groupId })
+      .populate("user_id", "username email")
+      .select("user_id role"); // chọn cả role để trả về loại thành viên
 
-    res.status(200).json({ message: "Danh sách thành viên", members });
+    // Tùy chỉnh lại dữ liệu trả về (gộp thông tin user và role)
+    const formattedMembers = members.map(member => ({
+      _id: member.user_id._id,
+      username: member.user_id.username,
+      email: member.user_id.email,
+      role: member.role,
+    }));
+
+    res.status(200).json({ message: "Danh sách thành viên", groupId: groupId ,members: formattedMembers });
   } catch (err) {
     res.status(500).json({ message: err.message });
   }
 };
 
+
 // Xóa nhóm
 // Hàm này sẽ được gọi khi người dùng muốn xóa nhóm
 const deleteGroup = async (req, res) => {
   try {
-    const { groupId } = req.body;
+    const groupId = req.params.groupId;
 
     // Xóa nhóm
     await Group.findByIdAndDelete(groupId);
@@ -171,7 +268,6 @@ const deleteGroup = async (req, res) => {
   }
 };
 
-
 export {
   createPrivateGroup,
   createGroupChat,
@@ -179,4 +275,5 @@ export {
   getUserGroups,
   getGroupMembers,
   deleteGroup,
+  getJoinedGroups,
 };
